Move tests for scanners to own submodule
