+++ CryptoAnalysis/.gitignore
@@ -1,1 +1,3 @@
 src/main/resources/**/*.crysl
+src/test/resources/rules/*
+!src/test/resources/rules/.gitkeep
