+++ HeadlessAndroidScanner/.gitignore
@@ -1,3 +1,3 @@
 src/main/resources/**/*.crysl
-src/test/resources/platforms/*
-!src/test/resources/platforms/.gitkeep
+src/test/resources/rules/*
+!src/test/resources/rules/.gitkeep
