+++ HeadlessJavaScanner/.gitignore
@@ -1,1 +1,2 @@
 src/main/resources/**/*.crysl
+src/test/resources/*
