+++ ScannerTests/.gitignore
@@ -1,0 +1,4 @@
+src/test/resources/rules/*
+!src/test/resources/rules/.gitkeep
+src/test/resources/platforms/*
+!src/test/resources/platforms/.gitkeep
