+++ ScannerTests/src/test/java/android/AbstractAndroidTest.java
