+++ ScannerTests/src/test/java/scanner/BragaCryptoGoodUsesTest.java
@@ -1,4 +1,4 @@
-package headless;
+package scanner;
 
 import crypto.analysis.errors.CallToError;
 import crypto.analysis.errors.ConstraintError;
@@ -8,7 +8,7 @@ import crypto.analysis.errors.IncompleteOperationError;
 import crypto.analysis.errors.RequiredPredicateError;
 import crypto.analysis.errors.TypestateError;
 import de.fraunhofer.iem.scanner.HeadlessJavaScanner;
-import headless.FindingsType.FalsePositives;
+import scanner.FindingsType.FalsePositives;
 import org.junit.Ignore;
 import org.junit.Test;
