+++ ScannerTests/src/test/java/scanner/CogniCryptGeneratedCodeTest.java
@@ -1,4 +1,4 @@
-package headless;
+package scanner;
 
 import crypto.analysis.errors.CallToError;
 import crypto.analysis.errors.HardCodedError;
