+++ ScannerTests/src/test/java/scanner/ReportFormatTest.java
@@ -1,4 +1,4 @@
-package headless;
+package scanner;
 
 import crypto.reporting.Reporter;
 import de.fraunhofer.iem.scanner.HeadlessJavaScanner;
