+++ ScannerTests/src/test/java/scanner/TLSRuleTest.java
@@ -1,10 +1,10 @@
-package headless;
-
-import java.io.File;
+package scanner;
 
 import de.fraunhofer.iem.scanner.HeadlessJavaScanner;
 import org.junit.Ignore;
 import org.junit.Test;
+
+import java.io.File;
 
 public class TLSRuleTest extends AbstractHeadlessTest{
 	@Ignore
