+++ ScannerTests/src/test/resources/apk/AndroidXAppCompatActivityCallbackDebug.apk
