+++ ScannerTests/src/test/resources/apk/Issue268.apk
