+++ ScannerTests/src/test/resources/apk/NoCallBackDebug.apk
