+++ ScannerTests/src/test/resources/apk/NoCallbackReleaseSigned.apk
