+++ ScannerTests/src/test/resources/apk/NoCallbackReleaseUnsigned.apk
