+++ ScannerTests/src/test/resources/apk/NormalActivityCallbackDebug.apk
