+++ ScannerTests/src/test/resources/apk/NormalAppCompatActivityCallbackDebug.apk
